requirements.txt
fusepy
requests
